# cmd/quiz/docs/swagger.yaml
basePath: /api
definitions:
  app_pkg_article_domain_entity.ArticleDTO:
    properties:
      categoryIds:
        items:
          type: integer
        type: array
      content:
        type: string
      publishedAt:
        type: string
      slug:
        type: string
      tagIds:
        items:
          type: integer
        type: array
      title:
        type: string
    required:
    - content
    - title
    type: object
...
  fiber.Map:
    additionalProperties: true
    type: object
...
  validation.ValidationError:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
host: 157.245.61.194:8082
info:
  contact:
    email: support@nosmo.com
    name: API Support
    url: http://www.nosmo.com/support
  description: This is a Quiz application API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Quiz API
  version: "1.0"
paths:
...
  /articles:
    get:
      consumes:
      - application/json
      description: Get all articles with pagination and filtering options
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Filter by category ID
        in: query
        name: categoryId
        type: integer
      - description: Filter by published status
        in: query
        name: published
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Create a new article with the provided information
      parameters:
      - description: Article information
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/app_pkg_article_domain_entity.ArticleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new article
      tags:
      - articles
  /articles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an article by its ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete an article
      tags:
      - articles
    get:
      consumes:
      - application/json
      description: Get details of a specific article by its ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get article by ID
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Update an article with the provided information
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated article information
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/app_pkg_article_domain_entity.ArticleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update an existing article
      tags:
      - articles
  /articles/{id}/publish:
    post:
      consumes:
      - application/json
      description: Change the status of an article to published
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Publish an article
      tags:
      - articles
  /articles/{id}/unpublish:
    post:
      consumes:
      - application/json
      description: Change the status of an article to unpublished
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Unpublish an article
      tags:
      - articles
  /articles/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get details of a specific article by its slug
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get article by slug
      tags:
      - articles
...
  /tags/article/{articleId}:
    get:
      consumes:
      - application/json
      description: Get all tags associated with a specific article
      parameters:
      - description: Article ID
        in: path
        name: articleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validation.ValidationError'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get tags by article
      tags:
      - tags
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT authorization with Bearer prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"